+++ persistent-test/src/Init.hs
@@ -1,6 +1,8 @@
-{-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE CPP #-}
+{-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE GADTs #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
@@ -42,14 +44,16 @@ module Init (
   , liftA2
   , changeBackend
   , Proxy(..)
+  , UUID(..)
+  , sqlSettingsUuid
   ) where
 
 #if !MIN_VERSION_monad_logger(0,3,30)
 -- Needed for GHC versions 7.10.3. Can drop when we drop support for GHC
 -- 7.10.3
+import qualified Control.Monad.Fail as MonadFail
 import Control.Monad.IO.Class
 import Control.Monad.Logger
-import qualified Control.Monad.Fail as MonadFail
 #endif
 
 -- needed for backwards compatibility
@@ -64,21 +68,35 @@ import Control.Monad.Trans.Resource.Internal
 -- re-exports
 import Control.Applicative (liftA2, (<|>))
 import Control.Exception (SomeException)
-import Control.Monad (void, replicateM, liftM, when, forM_)
+import Control.Monad (forM_, liftM, replicateM, void, when)
 import Control.Monad.Fail (MonadFail)
 import Control.Monad.Reader
-import Data.Char (generalCategory, GeneralCategory(..))
-import Data.Fixed (Pico,Micro)
+import Data.Char (GeneralCategory(..), generalCategory)
+import Data.Fixed (Micro, Pico)
+import Data.Proxy
 import qualified Data.Text as T
 import Data.Time
 import Test.Hspec
 import Test.QuickCheck.Instances ()
-import Data.Proxy
 
-import Database.Persist.TH (mkPersist, mkMigrate, share, sqlSettings, persistLowerCase, persistUpperCase, MkPersistSettings(..))
+import Data.Aeson (FromJSON, ToJSON, Value(..))
+import qualified Data.Text.Encoding as TE
+import Database.Persist.ImplicitIdDef (mkImplicitIdDef)
+import Database.Persist.TH
+       ( MkPersistSettings(..)
+       , mkMigrate
+       , mkPersist
+       , persistLowerCase
+       , persistUpperCase
+       , setImplicitIdDef
+       , share
+       , sqlSettings
+       )
+import Web.Internal.HttpApiData
+import Web.PathPieces
 
 -- testing
-import Test.HUnit ((@?=),(@=?), Assertion, assertFailure, assertBool)
+import Test.HUnit (Assertion, assertBool, assertFailure, (@=?), (@?=))
 import Test.QuickCheck
 
 import Control.Monad (unless, (>=>))
@@ -247,3 +265,34 @@ instance MonadBaseControl b m => MonadBaseControl b (ResourceT m) where
              f $ runInBase . (\(ResourceT r) -> r reader')
      restoreM = ResourceT . const . restoreM
 #endif
+
+-- * For implicit ID spec
+
+newtype UUID = UUID { unUUID :: Text }
+    deriving stock
+        (Show, Eq, Ord, Read)
+    deriving newtype
+        (ToJSON, FromJSON, FromHttpApiData, ToHttpApiData, PathPiece)
+
+instance PersistFieldSql UUID where
+    sqlType _ = SqlOther "UUID"
+
+instance PersistField UUID where
+    toPersistValue (UUID txt) =
+        PersistLiteral_ Escaped (TE.encodeUtf8 txt)
+    fromPersistValue pv =
+        case pv of
+            PersistLiteral_ Escaped bs ->
+                Right $ UUID (TE.decodeUtf8 bs)
+            _ ->
+                Left "Nope"
+
+sqlSettingsUuid :: Text -> MkPersistSettings
+sqlSettingsUuid defExpr =
+    let
+        uuidDef =
+           mkImplicitIdDef @UUID defExpr
+        settings =
+            setImplicitIdDef uuidDef sqlSettings
+     in
+        settings